requirements.txt
aiohttp==3.7.2
requests==2.23.0
progressbar2==3.53.1